requirements.txt
# ── Core Framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Deep Learning ────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
timm==0.9.16               # EfficientNet / modern backbones

# ── Computer Vision ──────────────────────────────────────────────
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0

# ── ML Utilities ─────────────────────────────────────────────────
scikit-learn==1.4.2
grad-cam==1.5.0            # pytorch-grad-cam

# ── Metadata / EXIF ──────────────────────────────────────────────
piexif==1.1.3
ExifRead==3.0.0

# ── Async / HTTP ─────────────────────────────────────────────────
aiofiles==23.2.1
httpx==0.27.0

# ── Validation & Config ──────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Logging ──────────────────────────────────────────────────────
structlog==24.1.0

# ── Dev / Test ───────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
